@@ -14,7 +14,7 @@
     "eslint"
   ],
   "dependencies": {
-    "eslint": "^8.57.0",
+    "eslint": "^9.0.0",
     "eslint-plugin-office-addins": "^4.0.1"
   },
   "peerDependencies": {
@@ -22,9 +22,9 @@
   },
   "devDependencies": {
     "@types/node": "^14.17.2",
-    "concurrently": "^6.2.2",
+    "concurrently": "^9.0.0",
     "office-addin-lint": "^3.0.1",
-    "rimraf": "^2.7.1",
+    "rimraf": "^6.0.1",
     "typescript": "^4.7.4"
   },
   "repository": {
