@@ -16,24 +16,24 @@
     "test": "jest"
   },
   "dependencies": {
-    "eslint": "^8.57.0",
+    "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-native": "^4.1.0",
+    "eslint-plugin-react-native": "^5.0.0",
     "office-addin-prettier-config": "^2.0.0",
     "prettier": "^3.2.5",
     "requireindex": "~1.2.0",
     "typescript": "^5.4.3",
-    "typescript-eslint": "^7.5.0"
+    "typescript-eslint": "^8.0.0"
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
     "@types/node": "^20.12.3",
-    "@typescript-eslint/rule-tester": "^7.5.0",
-    "concurrently": "^8.2.2",
+    "@typescript-eslint/rule-tester": "^8.0.0",
+    "concurrently": "^9.0.0",
     "jest": "^29.7.0",
-    "rimraf": "^3.0.2",
+    "rimraf": "^6.0.1",
     "ts-jest": "^29.1.2"
   },
   "engines": {
