@@ -21,23 +21,23 @@
     "Office"
   ],
   "dependencies": {
-    "commander": "^6.2.1",
+    "commander": "^13.0.0",
     "node-fetch": "^2.6.1",
     "read-package-json-fast": "^2.0.2"
   },
   "devDependencies": {
-    "@types/mocha": "^8.2.2",
+    "@types/mocha": "^10.0.6",
     "@types/node": "^14.17.2",
     "@types/node-fetch": "^2.5.10",
-    "concurrently": "^6.2.2",
-    "eslint": "^8.57.0",
+    "concurrently": "^9.0.0",
+    "eslint": "^9.0.0",
     "eslint-plugin-office-addins": "^4.0.1",
     "eslint-plugin-prettier": "^3.4.0",
-    "mocha": "^9.1.1",
+    "mocha": "^11.0.1",
     "office-addin-prettier-config": "^2.0.0",
     "office-addin-usage-data": "^2.0.1",
     "prettier": "^2.1.2",
-    "rimraf": "^3.0.2",
+    "rimraf": "^6.0.1",
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4",
     "typescript-eslint": "^8.4.0"
