@@ -24,8 +24,8 @@
     "Office Add-in"
   ],
   "dependencies": {
-    "commander": "^6.2.0",
-    "eslint": "^8.57.0",
+    "commander": "^13.0.0",
+    "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-office-addins": "^4.0.1",
     "eslint-plugin-prettier": "^5.1.3",
@@ -40,9 +40,9 @@
     "@types/node": "^20.12.4",
     "@types/node-fetch": "^2.6.11",
     "assert": "^2.1.0",
-    "concurrently": "^8.2.2",
-    "mocha": "^10.4.0",
-    "rimraf": "^5.0.5",
+    "concurrently": "^9.0.0",
+    "mocha": "^11.0.0",
+    "rimraf": "^6.0.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.3"
   },
