@@ -22,20 +22,20 @@
   ],
   "dependencies": {
     "applicationinsights": "^1.7.3",
-    "commander": "^6.2.0",
+    "commander": "^13.0.0",
     "readline-sync": "^1.4.9",
     "uuid": "8.3.2"
   },
   "devDependencies": {
-    "@types/mocha": "^5.2.7",
+    "@types/mocha": "^10.0.6",
     "@types/node": "^14.17.2",
-    "concurrently": "^6.0.2",
-    "eslint": "^8.57.0",
+    "concurrently": "^9.0.0",
+    "eslint": "^9.0.0",
     "eslint-plugin-office-addins": "^4.0.1",
     "eslint-plugin-prettier": "^5.2.1",
-    "mocha": "^9.1.1",
+    "mocha": "^11.0.1",
     "office-addin-prettier-config": "^2.0.0",
-    "rimraf": "^3.0.2",
+    "rimraf": "^6.0.1",
     "sinon": "^19.0.2",
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4",
