@@ -26,6 +26,7 @@ export default ESLintUtils.RuleCreator(
     schema: [],
   },
   create: function (context: any) {
+    const sourceCode = context.sourceCode ?? context.getSourceCode();
     type VariableProperty = {
       variable: string;
       property: string;
@@ -116,8 +117,11 @@ export default ESLintUtils.RuleCreator(
     }
 
     return {
-      Program() {
-        apiReferences = findOfficeApiReferences(context.getScope());
+      Program(node) {
+        const scope = sourceCode.getScope
+                    ? sourceCode.getScope(node)
+                    : context.getScope();
+        apiReferences = findOfficeApiReferences(scope);
         apiReferences.sort((left, right) => {
           return (
             left.reference.identifier.range[1] -
