@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import commander from "commander";
+import { OptionValues } from "commander";
 import fs from "fs";
 import { parseNumber, getPackageJsonScript } from "office-addin-cli";
 import { logErrorMessage } from "office-addin-usage-data";
@@ -43,7 +43,7 @@ function parseDevServerPort(optionValue: any): number | undefined {
   return devServerPort;
 }
 
-export async function start(manifestPath: string, platform: string | undefined, command: commander.Command) {
+export async function start(manifestPath: string, platform: string | undefined, options: OptionValues) {
   try {
     const appPlatformToDebug: Platform | undefined = parsePlatform(
       platform || process.env.npm_package_config_app_platform_to_debug || Platform.Win32
@@ -51,25 +51,25 @@ export async function start(manifestPath: string, platform: string | undefined, 
     const appTypeToDebug: AppType | undefined = devSettings.parseAppType(
       appPlatformToDebug || process.env.npm_package_config_app_type_to_debug || AppType.Desktop
     );
-    const appToDebug: string | undefined = command.app || process.env.npm_package_config_app_to_debug;
+    const appToDebug: string | undefined = options.app || process.env.npm_package_config_app_to_debug;
     const app: OfficeApp | undefined = appToDebug ? parseOfficeApp(appToDebug) : undefined;
-    const dev: boolean = command.prod ? false : true;
-    const debuggingMethod = parseDebuggingMethod(command.debugMethod);
-    const devServer: string | undefined = command.devServer || (await getPackageJsonScript("dev-server"));
-    const devServerPort = parseDevServerPort(command.devServerPort || process.env.npm_package_config_dev_server_port);
-    const document: string | undefined = command.document || process.env.npm_package_config_document;
-    const enableDebugging: boolean = command.debug;
-    const enableLiveReload: boolean = command.liveReload === true;
-    const enableSideload: boolean = command.sideload !== false; // enable if true or undefined; only disable if false
-    const openDevTools: boolean = command.devTools === true;
-    const packager: string | undefined = command.packager || (await getPackageJsonScript("packager"));
-    const packagerHost: string | undefined = command.PackagerHost || process.env.npm_package_config_packager_host;
-    const packagerPort: string | undefined = command.PackagerPort || process.env.npm_package_config_packager_port;
+    const dev: boolean = options.prod ? false : true;
+    const debuggingMethod = parseDebuggingMethod(options.debugMethod);
+    const devServer: string | undefined = options.devServer || (await getPackageJsonScript("dev-server"));
+    const devServerPort = parseDevServerPort(options.devServerPort || process.env.npm_package_config_dev_server_port);
+    const document: string | undefined = options.document || process.env.npm_package_config_document;
+    const enableDebugging: boolean = options.debug;
+    const enableLiveReload: boolean = options.liveReload === true;
+    const enableSideload: boolean = options.sideload !== false; // enable if true or undefined; only disable if false
+    const openDevTools: boolean = options.devTools === true;
+    const packager: string | undefined = options.packager || (await getPackageJsonScript("packager"));
+    const packagerHost: string | undefined = options.PackagerHost || process.env.npm_package_config_packager_host;
+    const packagerPort: string | undefined = options.PackagerPort || process.env.npm_package_config_packager_port;
     const sourceBundleUrlComponents = new devSettings.SourceBundleUrlComponents(
-      command.sourceBundleUrlHost,
-      command.sourceBundleUrlPort,
-      command.sourceBundleUrlPath,
-      command.sourceBundleUrlExtension
+      options.sourceBundleUrlHost,
+      options.sourceBundleUrlPort,
+      options.sourceBundleUrlPath,
+      options.sourceBundleUrlExtension
     );
 
     if (appPlatformToDebug === undefined) {
@@ -112,12 +112,12 @@ export async function start(manifestPath: string, platform: string | undefined, 
   }
 }
 
-export async function stop(manifestPath: string, platform: string | undefined, command: commander.Command) {
+export async function stop(manifestPath: string, platform: string | undefined, options: OptionValues) {
   try {
     const appPlatformToDebug: Platform | undefined = parsePlatform(
       platform || process.env.npm_package_config_app_plaform_to_debug || Platform.Win32
     );
-    const dev: boolean = command.prod ? false : true;
+    const dev: boolean = options.prod ? false : true;
 
     if (manifestPath === "" && appPlatformToDebug !== undefined) {
       manifestPath = determineManifestPath(appPlatformToDebug, dev);
