@@ -6,14 +6,17 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-import commander from "commander";
+import { Command } from "commander";
 import { run } from "./debugger";
 
 /* global process */
+
+const commander = new Command();
 
 commander
   .option("-h, --host <host>", "The hostname where the packager is running.")
   .option("-p, --port <port>", "The port where the packager is running.")
   .parse(process.argv);
 
-run(commander.host, commander.port);
+const options = commander.opts();
+run(options.host, options.port);
