@@ -1,12 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import commander from "commander";
+import { Command } from "commander";
 import { logErrorMessage } from "office-addin-usage-data";
 import * as commands from "./commands";
 import { defaultPort } from "./testServer";
 
 /* global process */
+
+const commander = new Command();
 
 commander.name("office-addin-test-server");
 commander.version(process.env.npm_package_version || "(version not available)");
